Agregando implementacion de CuentaCorriente y CuentaAhorros Correcciones en la el codigo Eliminacion de redundancia en setter y getters Persona.h invalido de momento, por corregir

+++ AplicacionBancoCuenta/Banco.h
@@ -30,7 +30,7 @@ public:
 	void buscarCuentasPorFecha(const std::string& fecha) const;
 	void buscarCuentasPorCriterio(); // Busca cuentas por criterio
 	void realizarTransferencia(); // Metodo para realizar transferencias entre cuentas
-	std::string formatearConComas(int valorEnCentavos)const;
+	std::string formatearConComas(double valorEnCentavos)const;
 
 	// Obtiene la ruta del escritorio del usuario
 	std::string obtenerRutaEscritorio() const;
